move dimension to common
