+++ include/stencil-composition/structured_grids/accessor_cxx11.hpp
@@ -37,7 +37,7 @@
 #pragma once
 #include "../accessor_base.hpp"
 #include "../arg.hpp"
-#include "../dimension.hpp"
+#include "../../common/dimension.hpp"
 #include "../../common/generic_metafunctions/all_integrals.hpp"
 #include "../../common/generic_metafunctions/static_if.hpp"
